backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = ""
authors = [{ name = "captcha781", email = "bhuvanesh19112001@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11.0,<4.0"
dependencies = [
  "fastapi (>=0.115.13,<0.116.0)",
  "uvicorn (>=0.34.3,<0.35.0)",
  "python-decouple (>=3.8,<4.0)",
  "beanie (>=1.30.0,<2.0.0)",
  "fastapi-jwt (>=0.3.0,<0.4.0)",
  "python-jose (>=3.5.0,<4.0.0)",
  "bcrypt (>=4.3.0,<5.0.0)",
  "cryptography (>=45.0.4,<46.0.0)",
  "python-multipart (>=0.0.20,<0.0.21)",
  "astrapy (>=2.0.1,<3.0.0)",
  "numpy (>=2.3.0,<3.0.0)",
  "langchain-google-genai (>=2.1.5,<3.0.0)",
  "langgraph (>=0.4.8,<0.5.0)",
  "chardet (>=5.2.0,<6.0.0)",
  "pillow (>=11.2.1,<12.0.0)",
  "pytesseract (>=0.3.13,<0.4.0)",
  "pymupdf (>=1.26.1,<2.0.0)",
]

[tool.poetry.scripts]
server = "backend.server:start_server"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']
exclude = '''
/(
    \.venv
  | \.git
  | \.mypy_cache
  | \.tox
  | \.pytest_cache
  | build
  | dist
)/
'''
